% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{.dmvt_arma}
\alias{.dmvt_arma}
\title{Multivariate Student density function}
\usage{
.dmvt_arma(x, mu, sigma, df, logd = FALSE)
}
\arguments{
\item{x}{matrix of observations}

\item{mu}{location vector}

\item{sigma}{positive definite scale matrix}

\item{df}{degrees of freedom}

\item{logd}{logical; whether log-density should be returned (default to \code{FALSE})}
}
\value{
density or log-density of the \code{nrow(x)} sample
}
\description{
This function returns the log-density for a multivariate Student distribution.
The data must be imputed as a matrix, using e.g., \code{as.matrix}, with each row
representing an observation.
}
\keyword{internal}
