% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmvnorm.R
\name{dtmvnorm}
\alias{dtmvnorm}
\title{Density function for the truncated multivariate normal distribution}
\usage{
dtmvnorm(x, mu, sigma, lb, ub, log = FALSE, type = c("mc", "qmc"), B = 10000)
}
\description{
This function returns the (log)-density of a matrix \code{x} of observations lying in the interval [\code{lb}, \code{ub}].
}
\seealso{
\code{\link{tmvnorm}}
}
\keyword{internal}
