% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{lnNpr}
\alias{lnNpr}
\title{Calculate log of Gaussian distribution function accurately}
\usage{
lnNpr(a, b, check = TRUE)
}
\arguments{
\item{a}{vector of lower bound}

\item{b}{vector of upper bound}

\item{check}{logical; should checks be performed?}
}
\value{
a vector of log probability.
}
\description{
This function returns the probability of a standard Gaussian variate between 
the interval \code{a} and \code{b}, avoiding numerical overflow. The function is vectorized
and is meant to be used only internally by the package \code{TruncatedNormal}.
}
\keyword{internal}
