% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{mroz}
\alias{mroz}
\title{Women wage dataset from Mroz (1987)}
\format{
a data frame containing the following variables:
\itemize{
\item \code{whrs}: hours of work
\item \code{kidslt6}: number of children aged 5 and below years old in household
\item \code{kidsge6}: number of children between age of 6 and 18 in household
\item \code{age}: age (in years)
\item \code{educ}: number of years in school
\item \code{hearn}: hourly earnings
\item \code{exp}: years of previous labor market experience
}
}
\source{
W. Greene's website, accessed 17.12.2019 at <http://www.stern.nyu.edu/~wgreene/Text/Edition7/TableF5-1.csv>.
}
\description{
The data are from the Panel Study of Income Dynamics (PSID) longitudinal study, 1976 wave.
They give the number of work hours of married women along with socio-economic variables and the number of children.
}
\references{
T. A. Mroz, 1987. \emph{The Sensitivity of an Empirical Model of Married Women's Hours of Work to Economic and Statistical Assumptions}, Econometrica, \bold{55}(4), pp. 765-799
}
\seealso{
\code{\link[carData]{Mroz}}
}
