% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{Phinv}
\alias{Phinv}
\title{Quantile of truncated Gaussian}
\usage{
Phinv(p, l, u)
}
\arguments{
\item{p}{vector of probabilities}

\item{l}{\code{d} vector of lower bounds}

\item{u}{\code{d} vector of upper bounds}
}
\value{
vector of quantiles
}
\description{
The function compute the \code{p}th quantile associated to the truncated standard Gaussian 
variate on the interval (\code{l},\code{u}).
}
\keyword{internal}
