% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmvnorm.R
\name{ptmvnorm}
\alias{ptmvnorm}
\title{Cumulative distribution function of the truncated multivariate normal distribution.}
\usage{
ptmvnorm(q, mu, sigma, lb, ub, log = FALSE, type = c("mc", "qmc"), B = 10000)
}
\description{
This function returns the (log)-distribution function of a matrix \code{q} of observations lying in the interval [\code{lb}, \code{ub}].
}
\seealso{
\code{\link{tmvnorm}}
}
\keyword{internal}
