% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tnorm.R
\name{qtnorm}
\alias{qtnorm}
\title{Quantile function using the inversion method}
\usage{
qtnorm(p, mu, sd, lb, ub)
}
\description{
If \code{p} is a matrix, the arguments are recycled.  If \code{mu} or \code{sd} are not specified they assume the default values of 0 and 1, respectively.
}
\seealso{
\code{\link{tnorm}}
}
\keyword{internal}
