% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmvnorm.R
\name{rtmvnorm}
\alias{rtmvnorm}
\title{Random number generator for the truncated multivariate normal distribution.}
\usage{
rtmvnorm(n, mu, sigma, lb, ub)
}
\description{
This function returns a matrix of draws from a multivariate normal distribution truncated on the interval [\code{lb}, \code{ub}].
}
\seealso{
\code{\link{tmvnorm}}
}
\keyword{internal}
