% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tnorm.R
\name{rtnorm}
\alias{rtnorm}
\title{Vectorized random number generation from the univariate truncated normal distribution}
\usage{
rtnorm(n, mu, sd, lb, ub, method = c("fast", "invtransfo"))
}
\description{
If \code{n} is 1, the function returns a vector rather than a matrix. If \code{mu} or \code{sd} are not specified they assume the default values of 0 and 1, respectively.
}
\seealso{
\code{\link{tnorm}}
}
\keyword{internal}
