% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tregress.R
\name{tregress}
\alias{tregress}
\title{Truncated student generator for Bayesian regression simulation}
\usage{
tregress(l, u, Sig, df, n)
}
\arguments{
\item{l}{lower bound for truncation (infinite values allowed)}

\item{u}{upper bound for truncation}

\item{Sig}{Covariance matrix}

\item{df}{degrees of freedom}

\item{n}{sample size}
}
\value{
list with components
\itemize{
\item{\code{R}: } \code{n} vector of scale
\item{\code{Z}: } a \code{d} by \code{n} matrix
} so that \eqn{\sqrt(\nu)Z/R} follows
a truncated Student distribution
}
\description{
Simulates \code{n} random vectors \eqn{X} exactly distributed
from the \code{d}-dimensional Student distribution with
\code{df=}\eqn{\nu} degrees of freedom, mean zero and scale matrix
\code{Sig}, conditional on \eqn{l<X<u},
}
\examples{
d <- 5
tregress(rep(-2, d), rep(2, d), df = 3, n = 10,
  Sig = diag(0.5, d) + matrix(1, d, d))
}
\references{
Z. I. Botev and P. L'Ecuyer (2015), Efficient probability estimation
and simulation of the truncated multivariate Student-t distribution,
Proceedings of the 2015 Winter Simulation Conference, pp. 380-391,
}
\author{
\code{Matlab} code by Zdravko Botev, \code{R} port by Leo Belzile
}
