% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{lupus}
\alias{lupus}
\title{Latent membranous Lupus Nephritis dataset}
\format{
a data frame with columns "response", "const", "x1" and "x2"
}
\source{
The data were transcribed from Table 1, page 22, of Dyk and Meng (2001).
}
\description{
The data represents two clinical measurements (covariates), which are used to predict the occurrence of latent membranous lupus nephritis. The dataset consists of measurements on 55 patients of which 18 have been diagnosed with latent membranous lupus.
}
\references{
D. A. van Dyk and X.-L. Meng (2001) \emph{The art of data augmentation (with discussion)}. Journal of Computational and  Graphical Statistics, volume 10, pages 1-50.
}
\seealso{
The dataset is used in the examples of \code{\link{mvrandn}}
}
