\name{TruncatedNormal-package}
\alias{TruncatedNormal}
\docType{package}
\title{
 Truncated Normal Distribution Toolbox
}
\description{
The routines include:
\itemize{
\item generator of \bold{independent and identically distributed} random vectors from the truncated univariate and multivariate distributions;
\item (Quasi-) Monte Carlo estimator and a \bold{deterministic upper bound}  of the cumulative distribution function of the multivariate normal and Student distributions;
\item algorithm for the accurate computation of the quantile function of the normal distribution in the extremes of its tails.
}
}

\author{Leo Belzile and Z. I. Botev, email: \email{botev@unsw.edu.au} and web page: \code{https://web.maths.unsw.edu.au/~zdravkobotev/}
}
\references{
\itemize{
\item Z. I. Botev (2017), \emph{The Normal Law Under Linear Restrictions:
Simulation and Estimation via Minimax Tilting}, Journal of the Royal
 Statistical Society, Series B, \bold{79} (1), pp. 1--24.
 \item Z. I. Botev and P. L'Ecuyer (2015), \emph{Efficient Estimation
 and Simulation of the Truncated Multivariate Student-t Distribution}, Proceedings of the 2015 Winter Simulation Conference,
   Huntington Beach, CA, USA
 \item Gibson G. J., Glasbey C. A., Elston D. A. (1994),
  \emph{Monte Carlo evaluation of multivariate normal integrals and sensitivity to variate ordering},
  In: Advances in Numerical Methods and Applications, pages 120--126
 }
}
